requirements.txt
streamlit==1.45.1
requests
biopython
pubchempy
py3Dmol
stmol
rdkit
Pillow
pandas
plotly
